pyproject.toml
[project]
name = "scrapping-imm"
version = "0.1.0"
description = "Système d'alerte immobilière avec Playwright et n8n"
authors = [
    {name = "EwenLive", email = "ewencroizier@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

dependencies = [
    "playwright==1.49.1",
    "playwright-stealth==1.0.6",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "fake-useragent==2.0.3",
    "setuptools (>=80.9.0,<81.0.0)"
]

# AJOUTE CES DEUX LIGNES ICI :
[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"